+++ client/src/app/_models/searchOptions.ts
@@ -1,0 +1,9 @@
+export interface SearchOptions {
+  query?: string,
+  majors?: string[],
+  GPA?: number,
+  npp?: number,
+  pageNum?: number,
+  posted?: Date,
+  deadline?: Date,
+};
